opkg.lua: drop ignore lists and use arguments for it instead
This removes usage of /etc/updater/opkg-ignore directory for skipping
custom feeds configured in opkg.

The was to skip feeds for those already added by custom configuration in
updater. That is because updater allows more security settings (or less
if you choose so) over just plain URL statement.

This removes original apporach but replaces it with possibility to
specify 'updater-ignore' argument.
Arguments are supported by opkg but mostly ignored (if not known) so
adding new one is not a problem.